+++ Tests/Unit/MSFT_xCluster.Tests.ps1
@@ -23,6 +23,7 @@ $TestEnvironment = Initialize-TestEnvironment `
 function Invoke-TestSetup
 {
     Import-Module -Name (Join-Path -Path (Join-Path -Path (Join-Path -Path (Join-Path -Path $script:moduleRoot -ChildPath 'Tests') -ChildPath 'Unit') -ChildPath 'Stubs') -ChildPath 'FailoverClusters.stubs.psm1') -Global -Force
+    Import-Module -Name (Join-Path -Path (Join-Path -Path (Join-Path -Path $script:moduleRoot -ChildPath 'Tests') -ChildPath 'TestHelpers') -ChildPath 'CommonTestHelper.psm1') -Global -Force
 }
 
 function Invoke-TestCleanup
@@ -110,7 +111,12 @@ try
 
                     Mock -CommandName Get-CimInstance -MockWith $mockGetCimInstance -ParameterFilter $mockGetCimInstance_ParameterFilter -Verifiable
 
+<<<<<<< HEAD
                     { Get-TargetResource @mockDefaultParameters } | Should Throw "Can't find machine's domain name"
+=======
+                    $mockCorrectErrorRecord = Get-InvalidOperationRecord -Message $script:localizedData.TargetNodeDomainMissing
+                    { Get-TargetResource @mockDefaultParameters } | Should -Throw $mockCorrectErrorRecord
+>>>>>>> Changes to xCluster
                 }
             }
 
@@ -122,7 +128,12 @@ try
                     Mock -CommandName Get-Cluster -Verifiable
                     Mock -CommandName Get-CimInstance -MockWith $mockGetCimInstance -ParameterFilter $mockGetCimInstance_ParameterFilter -Verifiable
 
+<<<<<<< HEAD
                     { Get-TargetResource @mockDefaultParameters } | Should Throw ("Can't find the cluster {0}" -f $mockClusterName)
+=======
+                    $mockCorrectErrorRecord = Get-ObjectNotFoundException -Message ($script:localizedData.ClusterNameNotFound -f $mockClusterName)
+                    { Get-TargetResource @mockDefaultParameters } | Should -Throw $mockCorrectErrorRecord
+>>>>>>> Changes to xCluster
                 }
             }
 
@@ -160,7 +171,12 @@ try
 
                     Mock -CommandName Get-CimInstance -MockWith $mockGetCimInstance -ParameterFilter $mockGetCimInstance_ParameterFilter -Verifiable
 
+<<<<<<< HEAD
                     { Set-TargetResource @mockDefaultParameters } | Should Throw "Can't find machine's domain name"
+=======
+                    $mockCorrectErrorRecord = Get-InvalidOperationRecord -Message $script:localizedData.TargetNodeDomainMissing
+                    { Set-TargetResource @mockDefaultParameters } | Should -Throw $mockCorrectErrorRecord
+>>>>>>> Changes to xCluster
                 }
             }
 
@@ -215,7 +231,12 @@ try
                     It 'Should throw the correct error message' {
                         Mock -CommandName Get-Cluster
 
+<<<<<<< HEAD
                         { Set-TargetResource @mockDefaultParameters } | Should Throw 'Cluster creation failed. Please verify output of ''Get-Cluster'' command'
+=======
+                        $mockCorrectErrorRecord = Get-InvalidOperationRecord -Message $script:localizedData.FailedCreatingCluster
+                        { Set-TargetResource @mockDefaultParameters } | Should -Throw $mockCorrectErrorRecord
+>>>>>>> Changes to xCluster
 
                         Assert-MockCalled -CommandName New-Cluster -Exactly -Times 1 -Scope It
                         Assert-MockCalled -CommandName Remove-ClusterNode -Exactly -Times 0 -Scope It
@@ -303,7 +324,8 @@ try
 
                     Mock -CommandName Get-CimInstance -MockWith $mockGetCimInstance -ParameterFilter $mockGetCimInstance_ParameterFilter -Verifiable
 
-                    { Test-TargetResource @mockDefaultParameters } | Should Throw "Can't find machine's domain name"
+                    $mockCorrectErrorRecord = Get-InvalidOperationRecord -Message $script:localizedData.TargetNodeDomainMissing
+                    { Test-TargetResource @mockDefaultParameters } | Should Throw $mockCorrectErrorRecord
                 }
             }